refactor: specification 패키지로 분리
